requirements.txt
aiogram==2.25
aiohttp==3.8.4
aioschedule==0.5.2
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
certifi==2022.12.7
charset-normalizer==3.1.0
Deprecated==1.2.14
distro==1.9.0
docopt==0.6.2
frozenlist==1.3.3
idna==3.4
lxml==5.1.0
magic-filter==1.0.9
multidict==6.0.4
netaddr==1.2.1
num2words==0.5.13
python-docx==1.1.0
python-dotenv==1.0.1
pytz==2022.7.1
requests==2.31.0
six==1.16.0
typing_extensions==4.10.0
urllib3==2.2.1
wrapt==1.16.0
yarl==1.8.2
yookassa==3.0.1
